+++ src/extension.ts
@@ -9,6 +9,11 @@ import { logger } from "./logger";
 import { MessageManager } from "./message";
 import { InputBoxMinibuffer } from "./minibuffer";
 
+// HACK: Currently there is no official type-safe way to handle
+//       the unsafe inputs such as the arguments of the extensions.
+// See: https://github.com/microsoft/TypeScript/issues/37700#issuecomment-940865298
+type Unreliable<T> = { [P in keyof T]?: Unreliable<T[P]> } | Array<Unreliable<T>> | undefined;
+
 export function activate(context: vscode.ExtensionContext): void {
   MessageManager.registerDispose(context);
   Configuration.registerDispose(context);
@@ -50,7 +55,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   function registerEmulatorCommand(
     commandName: string,
-    callback: (emulator: EmacsEmulator, ...args: any[]) => any,
+    callback: (emulator: EmacsEmulator, ...args: Unreliable<any>[]) => any,
     onNoEmulator?: (...args: any[]) => any
   ) {
     const disposable = vscode.commands.registerCommand(commandName, (...args) => {
@@ -73,16 +78,20 @@ export function activate(context: vscode.ExtensionContext): void {
     registerEmulatorCommand(
       "type",
       (emulator, args) => {
+        const text = (args as unknown as { text: string }).text; // XXX: The arguments of `type` is guaranteed to have this signature.
         // Capture typing characters for prefix argument functionality.
-        logger.debug(`[type command]\t args.text = "${args.text}"`);
+        logger.debug(`[type command]\t args.text = "${text}"`);
 
-        return emulator.type(args.text);
+        return emulator.type(text);
       },
       (args) => vscode.commands.executeCommand("default:type", args)
     );
   }
 
   registerEmulatorCommand("emacs-mcx.subsequentArgumentDigit", (emulator, args) => {
+    if (!Array.isArray(args)) {
+      return;
+    }
     const arg = args[0];
     if (typeof arg !== "number") {
       return;
@@ -91,6 +100,9 @@ export function activate(context: vscode.ExtensionContext): void {
   });
 
   registerEmulatorCommand("emacs-mcx.digitArgument", (emulator, args) => {
+    if (!Array.isArray(args)) {
+      return;
+    }
     const arg = args[0];
     if (typeof arg !== "number") {
       return;
@@ -99,6 +111,9 @@ export function activate(context: vscode.ExtensionContext): void {
   });
 
   registerEmulatorCommand("emacs-mcx.typeChar", (emulator, args) => {
+    if (!Array.isArray(args)) {
+      return;
+    }
     const arg = args[0];
     if (typeof arg !== "string") {
       return;
@@ -143,7 +158,10 @@ export function activate(context: vscode.ExtensionContext): void {
   registerEmulatorCommand("emacs-mcx.isearchExit", async (emulator, args) => {
     await emulator.runCommand("isearchExit");
 
-    const secondCommand = args?.then;
+    if (args == null || typeof args !== "object" || Array.isArray(args)) {
+      return;
+    }
+    const secondCommand = args.then;
     if (typeof secondCommand === "string") {
       await vscode.commands.executeCommand(secondCommand);
     }
@@ -320,12 +338,15 @@ export function activate(context: vscode.ExtensionContext): void {
   });
 
   registerEmulatorCommand("emacs-mcx.executeCommandWithPrefixArgument", (emulator, args) => {
-    if (typeof args === "object" && args != null) {
-      if ("command" in args) {
-        emulator.executeCommandWithPrefixArgument(args["command"], args["args"], args["prefixArgumentKey"]);
-      }
-    } else if (Array.isArray(args) && args.length >= 1) {
-      emulator.executeCommandWithPrefixArgument(args[0], args[1], args[2]);
+    if (typeof args !== "object" || args == null || Array.isArray(args)) {
+      return;
+    }
+
+    if (
+      typeof args?.command === "string" &&
+      (typeof args?.prefixArgumentKey === "string" || args?.prefixArgumentKey == null)
+    ) {
+      emulator.executeCommandWithPrefixArgument(args["command"], args["args"], args["prefixArgumentKey"]);
     }
   });
 }